Switch most data members of Config to be private.
When they collide (e.g. fabric 2's 'user' key) it's
super confusing, even if you theoretically wrote the
config framework in question :(

This limits that confusion surface to just methods,
which are also a lot easier to spot when you screw
up and get a method object back instead of actual
data.