Keep the last of the post tasks during deduplication.
The problem was that if both `a` and `b` tasks have same
`pre` and `post` hooks, the `a b` was extended and
deduplicated to `pre a post b`. But the post hook means that
it must be run after the original task.

With this commit there are 3 rules of deduplication:
- don't touch direct task calls (`a a b` must remain the same)
- keep only the first of the same pre-hooked tasks
- keep only the last of the same post-hooked tasks

Fixes #298

Signed-off-by: Máté Farkas <mate.farkas@balabit.com>