Turn bad/misleading test into enhancement of other test
Defaults are defaults and new should win. What should not get stomped
by the new class are higher levels like collection or defaults.
But that is simply regular hierarchy stuff and does not need its own
test...

However, it's worth adding a more realistic wrinkle to another test that
was more appropriate anyways.