Update library.rst
Following the instructions it was not obvious how the subclassed config should be added to the program definition. I needed to read the program API to find the relevant information: http://docs.pyinvoke.org/en/latest/api/program.html#module-invoke.program.

All examples previous to this section have been accurate and the examples contain all of the relevant information to get the example working.