Move default encoding guts out to reusable function, then use it in Result init
This avoids the corner case of having a None encoding, which under Python 2
causes encoding related asplosions. Not technically backwards compatible,
but I don't think anybody would have been very attached to those exceptions.

(In reality, that encoding is always the Runner encoding, which again, is always
using that refactored code snippet to set the value...)