Revert _part of_ 75c4d5e (re #552) to fix #660
See comments in #660 for rationale and testing.

tl;dr I believe this part of the fix for #552 was erroneous
and I cannot currently prove that it's necessary to fix
the main #552 bug; and I can prove that it is causing #660;
so, revert it is. Plus clarifying the tests that were
'fixed' by #552.