Realized OP's impl was overkill/caused bugs. Just set name.
Suspect maybe this changed in the years since it was filed,
or they just didn't test --list, which exposes the bug: cannot
display the transformed name of a collection.

But as it happens, the way we normalize everything to dict string key
lookups meant there was no need to special-case collections on lookup,
only needed the 'save' side of things.