Depend on PyYAML for yaml on python3
This removes the vendored version of PyYAML3 has been outdated for a
while. Over the last few years there have been a couple of high severity
CVEs against PyYAML that are unfixed in the vendored version.

By declaring a dependency on PyYAML we can make use of the upstream work
that goes into fixing those issues at the low cost of having it as the
only external dependency.