Preserve environment variables explicitly passed to Context.sudo.
sudo, by default, will not pass environment variables from the
context in which it is invoked to the context in creates to run
actual commands in. This is almost certainly the correct thing
to do but unfortunately also means that any environment variables
that are explicitly passed via the env= kwarg also get removed.

This change adds a --preserve-env= argument to the internal sudo
invocation that will ensure that arguments passed via the env=
kwarg are presented to the interior command while still ensuring
that the rest of the outer environment stays outside.