Update invoking-tasks.rst
Original output made it seem like the build task was being run again, but the pre-task of that (clean) was still being deduplicated. Running the example, in line with expectation, shows that the clean task is correctly also repeated if the --no-dedupe flag is set. This is purely a documentation fix.