Draft Idea For Pre Tasks To Be Context Aware
This commit introduces, in a draft state, the idea to allow tasks
that are called as a pre task for another task, to keep state
of this pre task assignment via the attribute `_is_pre_of`.

This attribute is later used when a `Call` object is created
from a `Task` instance that has said attribute set. A `Call`
created from a pre task, will add a reference to the task
that defines the current task as pre task, and said reference
will be available during the pre task's execution in the
context passed to the task.

The approach here suggested uses Pythonic patterns to avoid
changes to the class implementation as much as possible, but
rather compose the new functionality from invidivual, signle
purpose small components, such as a generic decorator, and
a descriptor component.

It can be very well the case that making a task more stateful
by tracking this pre assignment is not aligned with the design
if this library, in which case it would be better not to
add functionality that is not part of the long term view for
the library. I would be happy to contribute to this greaat
library in other ways.