fix: parser: detect '--help' if supplied to task
This fixes the invocation of a my_task which has mandatory/positional arguments:
```bash
> inv my_task --help
```

When detecting a missing positional arg, we check if the token is '--help/-h'.
If so, we store the task/context name to print help for it later and flag it the task/context to be completed without checks.

Signed-off-by: laurensmiers <laurens.miers@mind.be>