fix: program: print task-help if ParseError occurs
This will f.e. print the help of a task when it is not given its mandatory args:
```bash
> # Prints help if my_task expects mandatory positional arg
> inv my_task
```

If no name is available for the context, f.e. the core/None context,
we skip the specific task print since this means it was not a 'user' task which failed.

Signed-off-by: laurensmiers <laurens.miers@mind.be>